+++ src/Recyclarr.TrashLib/Services/CustomFormat/Models/ConflictingCustomFormat.cs
@@ -1,0 +1,6 @@
+namespace Recyclarr.TrashLib.Services.CustomFormat.Models;
+
+public record ConflictingCustomFormat(
+    ProcessedCustomFormatData GuideCf,
+    int ConflictingId
+);