+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Custom Formats: Updates that conflict with existing CFs in Sonarr/Radarr are now skipped and a
+  warning is printed.
+
 ## [4.1.0] - 2022-12-30
 
 ### Added